feat:added search functionality
